+++ packages/server/lib/cloud/api/index.ts
@@ -656,4 +656,6 @@ export default {
   },
 
   retryWithBackoff,
+
+  publicKeyVersion: PUBLIC_KEY_VERSION,
 }