+++ packages/server/test/support/fixtures/cloud/studio/test-studio.ts
@@ -1,0 +1,8 @@
+import type { AppStudioShape } from '@packages/types'
+import type { Router } from 'express'
+
+export class AppStudio implements AppStudioShape {
+  initializeRoutes (router: Router): void {
+
+  }
+}